Changed the controllers functionality
