+++ src/controllers/balance.ts
@@ -1,5 +1,9 @@
-import {BalanceRes, getBalanceService} from '../services/balance';
+import {getBalanceService} from '../services/balance';
 import {KinClient} from "@kinecosystem/kin-sdk-node";
+
+export type BalanceRes = {
+	balance: number
+}
 
 export type GetBalance = Request & {
 	params: {
@@ -12,5 +16,6 @@ export type GetBalance = Request & {
  * @returns {BalanceRes}
  */
 export async function getBalance(client: KinClient, address: string): Promise<BalanceRes> {
-	return await getBalanceService(client, address);
+	const balance = await getBalanceService(client, address);
+	return { balance: balance };
 }
