+++ src/controllers/payment.ts
@@ -1,5 +1,14 @@
 import {KinClient} from "@kinecosystem/kin-sdk-node";
-import {PaymentRes, paymentService} from "../services/payment";
+import {paymentService} from "../services/payment";
+import {InvalidTransactionError} from "../errors";
+
+export type PaymentRes = {
+	source: string,
+	destination: string,
+	amount: number,
+	memo?: string,
+	timestamp: number
+}
 
 export type GetPayment = Request & {
 	params: {
@@ -12,5 +21,18 @@ export type GetPayment = Request & {
  * @returns {PaymentRes}
  */
 export async function getPayment(client: KinClient, hash: string): Promise<PaymentRes> {
-	return await paymentService(client, hash);
+	const data = await paymentService(client, hash);
+	if (data.type !== 'PaymentTransaction') {
+		throw InvalidTransactionError();
+	}
+
+	// Convert from '2018-11-12T06:45:40Z' to unix timestamp
+	const timestamp = Date.parse(data.timestamp ? data.timestamp : new Date().toISOString());
+	return {
+		source: data.source,
+		destination: data.destination,
+		amount: data.amount,
+		memo: data.memo,
+		timestamp: timestamp
+	};
 }
