+++ src/services/pay.ts
@@ -1,15 +1,9 @@
 import {KinAccount, KinClient, LowBalanceError as KinLowBalanceError} from "@kinecosystem/kin-sdk-node";
 import {Pay} from "../controllers/pay";
-import {AccountNotFoundError} from "@kinecosystem/kin-sdk-node/scripts/bin/errors";
-import {DestinationDoesNotExistError, InvalidTransactionError, LowBalanceError} from "../errors";
+import {DestinationDoesNotExistError, LowBalanceError} from "../errors";
 
-export type PayRes = {
-	tx_id: string
-}
-
-export async function payService(client: KinClient, account: KinAccount, params: Pay): Promise<PayRes> {
+export async function payService(client: KinClient, account: KinAccount, params: Pay): Promise<string> {
 	const fee = await client.getMinimumFee();
-	let transactionId: string;
 	try {
 		const builder = await account.buildSendKin({
 			address: params.destination,
@@ -17,7 +11,7 @@ export async function payService(client: KinClient, account: KinAccount, params:
 			fee: fee,
 			memoText: params.memo
 		});
-		transactionId = await account.submitTransaction(builder);
+		return await account.submitTransaction(builder);
 	} catch (e) {
 		if (e.type === 'ResourceNotFoundError') {
 			throw DestinationDoesNotExistError(params.destination);
@@ -27,5 +21,4 @@ export async function payService(client: KinClient, account: KinAccount, params:
 			throw e;
 		}
 	}
-	return { tx_id: transactionId};
 }
